Make chagne to repeatString()
the stringWord variable was given an empty string as a value.
the for loop allowed the stringWord variable to be filled with a string argument each loop
for num amount of times.

return stringWord was placed after the for loop to return the value of the variable after changes made
during the loop statement ran.